% Generated by build script
\name{poly_events}
\alias{poly_events}
\title{Get Polymarket events (groups of markets)}
\usage{
poly_events()
}
\arguments{
\item{limit}{Number of events (default 25)}
}
\value{tibble: id, title, slug, end_date, volume, active}
\description{Get Polymarket events (groups of markets)}
