% Generated by build script
\name{poly_markets}
\alias{poly_markets}
\title{Get Polymarket markets}
\usage{
poly_markets()
}
\arguments{
\item{limit}{Number of markets (default 25)}
\item{active}{Show only active markets (default TRUE)}
}
\value{tibble of markets}
\description{Get Polymarket markets}
